refactor db initialization
