+++ src/dbstore/index.ts
@@ -1,16 +1,25 @@
 import { Config } from '../Config'
-import {
-  init,
-  close,
-  accountDatabase,
-  transactionDatabase,
-  cycleDatabase,
-  receiptDatabase,
-  originalTxDataDatabase,
-} from './sqlite3storage'
+import { Database } from 'sqlite3'
+import { close, createDB } from './sqlite3storage'
+
+export let cycleDatabase: Database
+export let accountDatabase: Database
+export let transactionDatabase: Database
+export let receiptDatabase: Database
+export let originalTxDataDatabase: Database
 
 export const initializeDB = async (config: Config): Promise<void> => {
-  await init(config)
+  accountDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.accountDB}`, 'Account')
+  cycleDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.cycleDB}`, 'Cycle')
+  transactionDatabase = await createDB(
+    `${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.transactionDB}`,
+    'Transaction'
+  )
+  receiptDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.receiptDB}`, 'Receipt')
+  originalTxDataDatabase = await createDB(
+    `${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.originalTxDataDB}`,
+    'OriginalTxData'
+  )
   // TODO: Check if the database have the required tables and they are in the same format as the original version in the archiver
 }
