+++ src/dbstore/originalTxsData.ts
@@ -1,5 +1,5 @@
 import * as db from './sqlite3storage'
-import { originalTxDataDatabase, extractValues, extractValuesFromArray } from './sqlite3storage'
+import { originalTxDataDatabase } from '.'
 import * as Logger from '../Logger'
 import { config } from '../Config'
 import { DeSerializeFromJsonString } from '../utils/serialization'
@@ -25,7 +25,7 @@ export async function insertOriginalTxData(OriginalTxData: OriginalTxData): Prom
   try {
     const fields = Object.keys(OriginalTxData).join(', ')
     const placeholders = Object.keys(OriginalTxData).fill('?').join(', ')
-    const values = extractValues(OriginalTxData)
+    const values = db.extractValues(OriginalTxData)
     if (!values || values.length === 0) {
       throw new Error(`No values extracted from OriginalTxData with txId ${OriginalTxData.txId}`)
     }
@@ -47,7 +47,7 @@ export async function bulkInsertOriginalTxsData(originalTxsData: OriginalTxData[
   try {
     const fields = Object.keys(originalTxsData[0]).join(', ')
     const placeholders = Object.keys(originalTxsData[0]).fill('?').join(', ')
-    const values = extractValuesFromArray(originalTxsData)
+    const values = db.extractValuesFromArray(originalTxsData)
     if (!values || values.length === 0) {
       throw new Error(
         `No values extracted from originalTxsData. Number of originalTxsData: ${originalTxsData.length}`
