+++ src/dbstore/sqlite3storage.ts
@@ -1,15 +1,10 @@
-import * as fs from 'fs'
-import * as path from 'path'
-import { Config } from '../Config'
 import { SerializeToJsonString } from '../utils/serialization'
-import { verbose, Database } from 'sqlite3'
+import { Database } from 'sqlite3'
 import { DBCycle, Cycle } from './cycles'
 import { Receipt, DBReceipt } from './receipts'
 import { OriginalTxData } from './originalTxsData'
 import { DBTransaction, Transaction } from './transactions'
 import { DBAccount, AccountCopy } from './accounts'
-
-const sqlite3 = verbose()
 
 export interface DBOriginalTxData {
   txId: string
@@ -31,27 +26,7 @@ type DBRecord =
   | DBAccount
   | AccountCopy
 
-export let cycleDatabase: Database
-export let accountDatabase: Database
-export let transactionDatabase: Database
-export let receiptDatabase: Database
-export let originalTxDataDatabase: Database
-
-export async function init(config: Config): Promise<void> {
-  accountDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.accountDB}`, 'Account')
-  cycleDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.cycleDB}`, 'Cycle')
-  transactionDatabase = await createDB(
-    `${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.transactionDB}`,
-    'Transaction'
-  )
-  receiptDatabase = await createDB(`${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.receiptDB}`, 'Receipt')
-  originalTxDataDatabase = await createDB(
-    `${config.ARCHIVER_DB_PATH}/${config.ARCHIVER_DATA.originalTxDataDB}`,
-    'OriginalTxData'
-  )
-}
-
-const createDB = async (dbPath: string, dbName: string): Promise<Database> => {
+export const createDB = async (dbPath: string, dbName: string): Promise<Database> => {
   console.log('CreateDB: dbName: ', dbName, 'dbPath: ', dbPath)
   const db = new Database(dbPath, (err) => {
     if (err) {
@@ -162,12 +137,4 @@ export function extractValuesFromArray(arr: DBRecord[]): (string | number | bool
   } catch (e) {
     console.log(e)
   }
-}
-
-// eslint-disable-next-line @typescript-eslint/no-unused-vars
-function createDirectories(pathname: string): void {
-  const __dirname = path.resolve()
-  pathname = pathname.replace(/^\.*\/|\/?[^/]+\.[a-z]+|\/$/g, '') // Remove leading directory markers, and remove ending /file-name.extension
-  // eslint-disable-next-line security/detect-non-literal-fs-filename
-  fs.mkdirSync(path.resolve(__dirname, pathname), { recursive: true })
 }
